Synchronization has been removed as an option for right now. - No promises it returns.
There are a lot of difficulties to get through with implementing this, as I am trying to keep this program as "portable" as possible.
Right now, it looks like adding synchronization features would require an initial "view" of the files to be made, and that view MUST be present in future runs to be used as a reference to know what was changed. Without it, we can't determine which files are "newer" when the last modified times are identical, but something else differs. Additionally, we don't know how to handle removed files or brand new files, which set of files contains the "newest" version when there is no file modification time to compare with.
I think that some kind of "file conflict" system could be put into place, but this would require a lot of work and thinking. - I am not interested in pursuing this right now. (I also don't personally use synchronization systems like that and have found that there are very good alternatives that exist, such as SyncThing (I think it just lacks hashing?))