--add-to-config now functions. It is not perfect, and needs tweaked further.
The leg work for reading from a configuration has also been done, but is untested and not fully implemented. A lot of the code was stolen from the --add-to-config stuff.

I am not particularly fond of my solution involving writing the altered config back to the file. As of now, there is the /potentially/ to lose the configuration with this.
This could potentially be remedied by creating a copy of the config, deleting the old version, then renaming the config.

Signed-off-by: Jadin Heaston <86203688+JadinHeaston@users.noreply.github.com>